+++ src/Angor/Avalonia/Angor.Projects/Domain/IInvestmentRepository.cs
@@ -1,7 +1,6 @@
-using Angor.Projects.Domain;
 using CSharpFunctionalExtensions;
 
-namespace Angor.Projects.Infrastructure.Interfaces;
+namespace Angor.Projects.Domain;
 
 public interface IInvestmentRepository
 {
