+++ src/Angor/Avalonia/Angor.Projects/Domain/InvestmentStatus.cs
@@ -1,0 +1,8 @@
+namespace Angor.Projects.Domain;
+
+public enum InvestmentStatus
+{
+    Pending,
+    Confirmed,
+    Failed
+}
