+++ src/Angor/Avalonia/Angor.Projects/ProjectServices.cs
@@ -5,7 +5,6 @@ using Angor.Projects.Infrastructure.Interfaces;
 using Angor.Shared;
 using Angor.Shared.Networks;
 using Angor.Shared.Services;
-using CSharpFunctionalExtensions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Serilog;
@@ -32,39 +31,5 @@ public class ProjectServices
         services.TryAddSingleton<INostrCommunicationFactory, NostrCommunicationFactory>();
         
         return services;
-    }
-}
-
-public class ProjectRepository : IProjectRepository
-{
-    public Task<Result<Project>> Get(ProjectId id)
-    {
-        throw new NotImplementedException();
-    }
-
-    public Task<Result> SaveAsync(Project project)
-    {
-        throw new NotImplementedException();
-    }
-}
-
-public class InvestmentRepository : IInvestmentRepository
-{
-    public Task<Result> Save(Investment investment)
-    {
-        throw new NotImplementedException();
-    }
-
-    public Task<Result<IEnumerable<Investment>>> Get(ProjectId projectId)
-    {
-        throw new NotImplementedException();
-    }
-}
-
-public class InvestmentService : IInvestmentService
-{
-    public Task<Result<string>> CreateInvestmentTransaction(string bitcoinAddress, string investorPubKey, long satoshiAmount, ModelFeeRate feeRate)
-    {
-        throw new NotImplementedException();
     }
 }
